addressed some comments
