+++ tests/trainer/fsdp2_context.py
@@ -11,6 +11,8 @@ SKIP_TEST = version.parse(torch.__version__) < version.parse('2.6.0')
 if not SKIP_TEST:
     # TODO (FSDP2) move this to top once we deprecate torch 2.5
     from composer.distributed import fsdp2
+    from composer.distributed import prepare_distributed
+    parallelize_model = prepare_distributed.parallelize_model
     prepare_fully_shard = fsdp2.prepare_fully_shard
     legalize_param_sharing_between_modules = fsdp2.legalize_param_sharing_between_modules
     get_standalone_and_tied_modules = fsdp2.get_standalone_and_tied_modules
@@ -18,6 +20,7 @@ if not SKIP_TEST:
     _generate_default_policy = fsdp2.generate_default_policy
     check_param_tying = fsdp2.check_param_tying
 else:
+    parallelize_model = lambda *args, **kwargs: None
     prepare_fully_shard = lambda *args, **kwargs: None
     legalize_param_sharing_between_modules = lambda *args, **kwargs: None
     get_standalone_and_tied_modules = lambda *args, **kwargs: ([], set())
